kobuki: update all 'time-reported' fields.
Previously the fields contained only approximate timestamps which also did not contain any timezone information. This caused issues when parsing the dates and passing them on to the time machine.

Some timestamps were off by as much as 2 hours, which could be due to GMT+1 and DST being active for the bug analyser.